Back to unsigned long, it breaks compilation on Solaris
git-svn-id: svn://anonsvn.kde.org/home/kde/trunk/kdesupport/taglib@1216598 283d02a7-25f6-0310-bc7c-ecb5cbfe19da

+++ taglib/mpeg/mpegheader.cpp
@@ -168,7 +168,7 @@ void MPEG::Header::parse(const ByteVector &data)
     return;
   }
 
-  std::bitset<32> flags(static_cast<unsigned long long>(data.toUInt()));
+  std::bitset<32> flags(static_cast<unsigned long>(data.toUInt()));
 
   // Check for the second byte's part of the MPEG synch
